requirements.txt
django
python-bitcoinrpc